requirements.txt
django==1.8.2
mysql-python==1.2.5
django-suit==0.2.18
poster==0.8.1
redis==2.10.5
django-redis==3.8.3 
#uwsgi==2.0.12
xpinyin==0.5.4
django-debug-toolbar==1.4
celery==3.1.23
django-celery==3.1.17
celery-with-redis==3.0
django_extensions==1.6.7
Pillow==3.2.0
easy_thumbnails==2.3
django-mysql==1.0.9
django-summernote==0.8.5
django-aliyun-oss2-storage==0.1.1
django-pagedown==0.1.1
djangorestframework==3.5.3
aliyun-python-sdk-push==3.0.3
pygal==2.3.1
qrcode==5.3
ipython==5.5.0
django-jsonfield==1.0.1
bitpay-py2==2.3.5
git+https://github.com/warner/python-ecdsa.git